libaries: rename AP_NMEA_Output to AP_NMEA
