+++ libraries/AP_NMEA/AP_NMEA_Output.cpp
