+++ libraries/AP_NMEA/examples/NMEA_Output/nmea_output.cpp
@@ -11,7 +11,7 @@
 #include <AP_Logger/AP_Logger.h>
 #include <AP_GPS/AP_GPS.h>
 #include <AP_Baro/AP_Baro.h>
-#include <AP_NMEA_Output/AP_NMEA_Output.h>
+#include <AP_NMEA/AP_NMEA_Output.h>
 #include <AP_SerialManager/AP_SerialManager.h>
 
 void setup();
